views: Create separate prometheus+bigquery components
